fix: move commitment_snapshot.rs out of actors and into database
