+++ crates/actors/src/lib.rs
@@ -7,7 +7,6 @@ pub mod block_validation;
 pub mod broadcast_mining_service;
 pub mod cache_service;
 pub mod chunk_migration_service;
-pub mod commitment_snapshot;
 pub mod ema_service;
 pub mod epoch_service;
 pub mod mempool_service;
@@ -21,6 +20,5 @@ pub mod validation_service;
 
 pub use addresses::*;
 pub use block_producer::*;
-pub use commitment_snapshot::*;
 pub use epoch_service::*;
 pub use storage_module_service::*;
