+++ crates/database/src/lib.rs
@@ -2,6 +2,7 @@
 //! database methods for reading and writing from the database as well as some
 //! database value types.
 pub mod block_index_data;
+pub mod commitment_snapshot;
 pub mod config;
 pub mod data_ledger;
 pub mod database;
@@ -26,6 +27,7 @@ pub mod submodule;
 pub mod tables;
 
 pub use block_index_data::*;
+pub use commitment_snapshot::*;
 pub use data_ledger::*;
 pub use database::*;
 pub use system_ledger::*;
